Refactor import paths and improve event listener management
