+++ src/js/game/levels/shapesLevel.js
@@ -1,6 +1,6 @@
 import { BaseLevel } from './baseLevel.js';
-import { GAME_CONFIG } from '../config/constants.js';
-import { LevelSettings } from '../config/gameSettings.js';
+import { GAME_CONFIG } from '../../config/constants.js';
+import { LevelSettings } from '../../config/gameSettings.js';
 
 export default class ShapesLevel extends BaseLevel {
   constructor(canvas, ctx, managers, helpers) {
